GardGameTemplate
I want to make a template project to make developing Cards game fastly,
so i classify the modules like this : hall, house, user, card, rule and
algorithm. the core of this project is how to make the rule and card
more extension. i define some framework to make it which we can do it
without developers, and we make a game with configure some xml. that,s
what i want to do.